refactor: move storage abstractions to their own namespace
