+++ Doppler.HtmlEditorApi/Controllers/TemplatesController.cs
@@ -7,7 +7,7 @@ using Doppler.HtmlEditorApi.DopplerSecurity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using Doppler.HtmlEditorApi.ApiModels;
-using Doppler.HtmlEditorApi.Infrastructure;
+using Doppler.HtmlEditorApi.Storage;
 
 
 namespace Doppler.HtmlEditorApi.Controllers
