--- Doppler.HtmlEditorApi/Infrastructure/IRepository.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Doppler.HtmlEditorApi.Infrastructure
-{
-    public interface IRepository
-    {
-        Task<ContentRow> GetCampaignModel(string accountName, int campaignId);
-        Task SaveCampaignContent(string accountName, ContentRow contentRow);
-    }
-}
