+++ Doppler.HtmlEditorApi/Infrastructure/Repository.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Dapper;
 using System.Text.Json;
 using System.Collections.Generic;
+using Doppler.HtmlEditorApi.Storage;
 
 namespace Doppler.HtmlEditorApi.Infrastructure
 {
