+++ Doppler.HtmlEditorApi/Storage/IRepository.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Doppler.HtmlEditorApi.Storage;
+
+public interface IRepository
+{
+    Task<ContentRow> GetCampaignModel(string accountName, int campaignId);
+    Task SaveCampaignContent(string accountName, ContentRow contentRow);
+}
